+++ lib/src/export.dart
@@ -87,6 +87,7 @@ export './options/option_payoff_chart.dart';
 export './volatility/volatility_curve_chart.dart';
 export './profile/market_profile_chart.dart';
 export './footprint/footprint_chart.dart';
+export './book/book_heatmap_chart.dart';
 export './gauge/gauge_chart.dart';
 export './treemap/treemap_chart.dart';
 export './treemap/treemap_data.dart';